pyproject.toml
[project]
name = "ninjacat-mcp"
version = "0.1.0"
description = "Model Context Protocol bridge that forwards OpenWebUI/n8n traffic to an in-house AI webhook."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "uvicorn>=0.30.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.2.0"
]

[project.scripts]
ninjacat-mcp = "ninjacat_mcp.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
